requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.2.10
astroid==2.4.2
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
beautifulsoup4==4.9.1
certifi==2020.4.5.2
cffi==1.14.4
channels==3.0.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
defusedxml==0.6.0
distlib==0.3.1
Django==3.0.8
django-allauth==0.44.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-js-asset==1.2.2
django-mptt==0.11.0
djangorestframework==3.11.1
docopt==0.6.2
filelock==3.0.12
gTTS==2.1.1
gTTS-token==1.1.3
gunicorn==20.0.4
hyperlink==21.0.0
idna==2.9
incremental==17.5.0
isort==5.6.4
Js2Py==0.70
lazy-object-proxy==1.4.3
mccabe==0.6.1
MouseInfo==0.1.3
oauthlib==3.1.0
packaging==20.4
Pillow==7.1.2
pipwin==0.5.0
playsound==1.2.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycparser==2.20
pygame==1.9.6
PyGetWindow==0.0.9
PyHamcrest==2.0.2
pyjsparser==2.7.1
PyJWT==1.7.1
pylint==2.6.0
PyMsgBox==1.0.9
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.1
PyPrind==2.11.2
PyRect==0.1.4
PyScreeze==0.1.26
pySmartDL==1.3.3
python3-openid==3.2.0
PyTweening==1.0.3
pytz==2020.1
pywhatkit==3.8
requests==2.24.0
requests-oauthlib==1.3.0
schedule==0.6.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
sqlparse==0.3.1
toml==0.10.2
twilio==6.50.1
txaio==20.12.1
tzlocal==2.1
urllib3==1.25.9
virtualenv==20.0.28
whitenoise==5.2.0
wikipedia==1.4.0
wrapt==1.12.1
youtube-dl==2021.1.8
zope.interface==5.2.0
